Drop unused methods from utils.dart
This file is not exported so there (hopefully) shouldn't be usage of any
of these methods outside of this package.

- Drop the methods that have no usage in this package
- Move the methods used only from tests to a library under test/
