+++ test/analysis_utils.dart
@@ -1,0 +1,111 @@
+import 'dart:async';
+import 'dart:io';
+
+import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/file_system/file_system.dart' hide File;
+import 'package:analyzer/file_system/physical_file_system.dart';
+import 'package:analyzer/source/package_map_resolver.dart';
+import 'package:analyzer/source/pub_package_map_provider.dart';
+import 'package:analyzer/src/dart/sdk/sdk.dart' show FolderBasedDartSdk;
+import 'package:analyzer/src/generated/engine.dart';
+import 'package:analyzer/src/generated/java_io.dart';
+import 'package:analyzer/src/generated/sdk.dart' show DartSdk;
+import 'package:analyzer/src/generated/source.dart';
+import 'package:analyzer/src/generated/source_io.dart';
+import 'package:cli_util/cli_util.dart' as cli;
+import 'package:path/path.dart' as p;
+
+/// [foundFiles] is the list of files to consider for the context.
+Future<AnalysisContext> getAnalysisContextForProjectPath(
+    String projectPath, List<String> foundFiles) async {
+  // TODO: fail more clearly if this...fails
+  var sdkPath = cli.getSdkDir().path;
+
+  JavaSystemIO.setProperty("com.google.dart.sdk", sdkPath);
+  var resourceProvider = PhysicalResourceProvider.INSTANCE;
+  DartSdk sdk = new FolderBasedDartSdk(
+      resourceProvider, resourceProvider.getFolder(sdkPath));
+
+  var packageResolver = _getPackageResolver(projectPath, sdk);
+
+  var resolvers = [
+    new DartUriResolver(sdk),
+    new ResourceUriResolver(PhysicalResourceProvider.INSTANCE),
+    packageResolver
+  ];
+
+  AnalysisEngine.instance.processRequiredPlugins();
+
+  var options = new AnalysisOptionsImpl()..analyzeFunctionBodies = false;
+
+  var context = AnalysisEngine.instance.createAnalysisContext()
+    ..analysisOptions = options
+    ..sourceFactory = new SourceFactory(resolvers);
+
+  // ensures all libraries defined by the set of files are resolved
+  _getLibraryElements(foundFiles, context).toList();
+
+  return context;
+}
+
+UriResolver _getPackageResolver(String projectPath, DartSdk sdk) {
+  var dotPackagesPath = p.join(projectPath, '.packages');
+
+  if (!FileSystemEntity.isFileSync(dotPackagesPath)) {
+    throw new StateError('A package configuration file was not found at the '
+        'expectetd location. $dotPackagesPath');
+  }
+
+  var pubPackageMapProvider =
+      new PubPackageMapProvider(PhysicalResourceProvider.INSTANCE, sdk);
+  var packageMapInfo = pubPackageMapProvider
+      .computePackageMap(PhysicalResourceProvider.INSTANCE.getResource('.'));
+  var packageMap = packageMapInfo.packageMap;
+  if (packageMap == null) {
+    throw new StateError('An error occurred getting the package map.');
+  }
+
+  return new PackageMapUriResolver(
+      PhysicalResourceProvider.INSTANCE, packageMap);
+}
+
+LibraryElement getLibraryElementForSourceFile(
+    AnalysisContext context, String sourcePath) {
+  Source source = new FileBasedSource(new JavaFile(sourcePath));
+
+  var libs = context.getLibrariesContaining(source);
+
+  if (libs.length > 1) {
+    throw new Exception("We don't support multiple libraries for a source.");
+  }
+
+  if (libs.isEmpty) {
+    return null;
+  }
+
+  var libSource = libs.single;
+
+  // using `getLibraryElement` because the library should already be computed
+  // If not, it's a bug in usage
+  return context.getLibraryElement(libSource);
+}
+
+// may return `null` if [path] doesn't refer to a library.
+/// [dartFiles] is a [Stream] of paths to [.dart] files.
+Iterable<LibraryElement> _getLibraryElements(
+        List<String> dartFiles, AnalysisContext context) =>
+    dartFiles
+        .map((path) => _getLibraryElement(path, context))
+        .where((lib) => lib != null);
+
+LibraryElement _getLibraryElement(String path, AnalysisContext context) {
+  Source source = new FileBasedSource(new JavaFile(path));
+  if (context.computeKindOf(source) == SourceKind.LIBRARY) {
+    return context.computeLibraryElement(source);
+  }
+  return null;
+}
+
+String getFileBasedSourcePath(FileBasedSource source) {
+  return p.fromUri(source.uri);
+}
