+++ test/find_libraries_test.dart
@@ -7,9 +7,9 @@ library source_gen.test.find_libraries;
 
 import 'package:analyzer/src/generated/engine.dart';
 import 'package:path/path.dart' as p;
-import 'package:source_gen/src/utils.dart';
 import 'package:test/test.dart';
 
+import 'analysis_utils.dart';
 import 'src/io.dart';
 import 'test_utils.dart';
